requirements.txt
Django
djangorestframework
psycopg2
django-debug-toolbar
python-dotenv